@danlark1 2021-06-02
> Awesome PR & analysis!
> 
> This is looking good to me! I've done some initial measurements and they look good. I will spend a bit more time measuring, then we should be good to merge this, with the change to make it C89 compliant.
> 
> Note that on my machine (i9-9900K) gcc-11 has some alignment issues that cause a ~10% regression. But, changing the alignment to `.p2align 4` fixes that. This is a known issue, see:
> 
> https://github.com/facebook/zstd/blob/18d02cbf2e0654de08093094f1a77cfd231f11d7/lib/decompress/zstd_decompress_block.c#L1128-L1175

Thanks, let me try out alignment for other gcc versions