@danlark1 2021-06-08
> > Thanks, let me try out alignment for other gcc versions
> 
> Not all Intel CPUs have this problem, it seems to only be the newer consumer CPUs. I haven't seen the same on Skylake server CPUs, but it may be present in some of the newer ones. You can use the perf commands referenced in the comments to see if you're running into the DSB issues.

After several rounds of evaluation, I agree gcc11 is worse, however, gcc9 and gcc10 are better with .palign 3. Adjusted to that