@terrelln 2021-06-08
> After several rounds of evaluation, I agree gcc11 is worse, however, gcc9 and gcc10 are better with .palign 3. Adjusted to that

Thanks for doing that! The alignment issues are super annoying, and make it very hard to get consistent measurements, since any small change to the decoder can cause a 10% drop in performance. The `.p2align` directives help stabilize it a bit, so at least changes outside the decoding loop generally don't affect its performance. But, it doesn't help when changing the core decoder logic.

I'll dedicate some time this week to measure performance across the various machines I have, then we can get this merged.