@animalize 2021-06-11
Just FYI, Intel released a Optimization Reference Manual yesterday:

https://www.phoronix.com/scan.php?page=news_item&px=Intel-ORM-Code-Samples