@terrelln 2021-06-15
Yeah, I ran some benchmarks on Friday & today.

I still see a consistent gcc-11 regression of ~4% on my i9-9900k (down from ~10%). But, I have confirmed the clang speedup.

It looks like gcc-11 is hitting the bad case of falling out of the DSB both before & after this patch. Comparing v148 and v150 on gcc-11 there is a 5% regression. And, I can't mess with the `.p2align` directives to fix it. I believe that the change in PR #2625 is the culprit. But, I'm trying to figure out if this performance loss is real, an if it affects just my i9-9900k or other machines, and if we can avoid a gcc performance loss. So far, I haven't been able to re-write the loop to regain the speed of v148 without reverting PR #2625.

I still have to measure on two more machines, both of which don't have alignment issues. I want to accept this PR because the changes make sense. And, performance parity with clang would be great. But, the fact that it isn't a pure win, and that there are already gcc regressions makes it tricky.

I'm collecting the results of my benchmarks, and I will share them here when I'm done. In the meantime I'm going to continue trying to fiddle with the code to make gcc happy.