@danlark1 2021-06-15
I confirm only gcc 11 is affected by 2-4%, all other versions of gcc stay the same for performance as noted in the description

gcc 11 laptop

![gcc_11_c89_optimized](https://user-images.githubusercontent.com/15173761/122062268-2d889f80-cde7-11eb-8d7a-4458ef28e822.png)

gcc 11 server

![gcc_11_server_optimized](https://user-images.githubusercontent.com/15173761/122063218-fa92db80-cde7-11eb-8e15-28c2a927038c.png)