@danlark1 2021-06-15
More graphs!

![gcc_11_server_comparison_optimized](https://user-images.githubusercontent.com/15173761/122066058-53fc0a00-cdea-11eb-864a-86d51807fb34.png)


gcc10 vs gcc11 comparison, it seems like gcc11 found new ways to optimize code even further. With that patch gcc11 codegen is the same as old gcc10 codegen