@danlark1 2021-06-20
I also was not able to remove 3-4% performance difference for gcc-11

I have several options:

1. Leave it as it is, as I showed the performance is the same as gcc-10 and we can try to optimize regressions from #2625 separately
2. Try to move old code under `__GNUC__ == 11` macro
3. Investigate more, however, first analysis showed it is much harder to reason about GCC changes and I believe the problem is somewhere in stack allocation

The patch is extremely valuable for other configurations and for us specifically, it would be unfortunate to abandon it