@danlark1 2021-06-29
I tried the patch, it does slightly better for clang and a little bit better for gcc-11, yet, I am still able to reproduce some regression for server gcc-11. Now it is 2% instead of 4%.

Anyway, I am good with this, thanks a lot, I appreciate the effort