@terrelln 2021-06-29
> I tried the patch, it does slightly better for clang and a little bit better for gcc-11, yet, I am still able to reproduce some regression for server gcc-11. Now it is 2% instead of 4%.

Glad to hear that you've found a slight improvement as well. I think we should merge this PR with the small gcc-11 regression, since it brings large benefits for clang. And hopefully we'll find a way to improve gcc down the line.

> I tried seqState->prevOffset[ll0] = seqState->prevOffset[0]; and it was worse

Yeah, that is what I tried first and found it to be worse as well. I didn't dig into why one works and the other doesn't, but my guess is that something (either the CPU or the compiler) doesn't like the potentially aliased stores, or it is just noise, and there is no good reason why it is worse.