@animalize 2021-07-07
The decompression speed of negative compression levels is reduced a bit.
AMD 3600X, MSVC 2019.

I only tested a special data, maybe the result is not universal.
Please also test the negative compression level in the future.

The `X` axis is the compression level, the `Y` axis is the speed (MiB/s). 

![canvas](https://user-images.githubusercontent.com/11244865/124699597-aaf28c00-df1d-11eb-9310-42843c6219ab.png)